Moved from half and fixed to float
Conversions when mixing usually more time intense that memory saved with fixed hand half.

+++ Demos/PaintAPlane/PaintAPlane-Display.shader
@@ -44,25 +44,25 @@ Shader "ShaderCastle/Demos/PaintAPlane/PaintAPlane-Display"
                 return o;
             }
 
-            half4 frag (v2f i) : SV_Target
+            float4 frag (v2f i) : SV_Target
             {
                 float3 worldNormal = normalize(i.worldNormal);
 
                 // sample the texture
-                half3 albedo = tex2D(_MainTex, i.uv);
+                float3 albedo = tex2D(_MainTex, i.uv);
                 saturate(albedo);
 
                 // Light
                 float3 _world_light_direction = normalize(_WorldSpaceLightPos0.xyz);
                 float3 lightColor = _LightColor0.rgb;
-                half3 ambientLight = UNITY_LIGHTMODEL_AMBIENT.rgb;
-                half3 NdotL = dot(worldNormal, _world_light_direction);
+                float3 ambientLight = UNITY_LIGHTMODEL_AMBIENT.rgb;
+                float3 NdotL = dot(worldNormal, _world_light_direction);
                 NdotL = saturate(NdotL);
-                half3 directLight = NdotL * lightColor.rgb;
+                float3 directLight = NdotL * lightColor.rgb;
 
-                half3 color = half3((directLight + ambientLight) * albedo);
+                float3 color = float3((directLight + ambientLight) * albedo);
 
-                return half4 (color, 1.0);
+                return float4 (color, 1.0);
             }
             ENDCG
         }
