+++ Implementations/BuildingPerTheme/TexturesAndColor/PlanarProjectionInsideManyLights.shader
@@ -29,7 +29,7 @@ Shader "ShaderCastle/Implementations/BuildingPerTheme/TexturesAndColor/PlanarPro
             float4 _albedo_ST;
             sampler2D _normalMap;
             sampler2D _arm;
-            half3 _ambientLightColor;
+            float3 _ambientLightColor;
             float3 _pointLightColor;
             float _pattern;
             float _intensity;
@@ -86,7 +86,7 @@ Shader "ShaderCastle/Implementations/BuildingPerTheme/TexturesAndColor/PlanarPro
                 return o;
             }
 
-            half4 frag (v2f i) : SV_Target {
+            float4 frag (v2f i) : SV_Target {
                 float3 worldPos = i.worldPos;
                 float3 objectOrigin = unity_ObjectToWorld._m03_m13_m23;
 
@@ -113,36 +113,36 @@ Shader "ShaderCastle/Implementations/BuildingPerTheme/TexturesAndColor/PlanarPro
 
                 float pointLightIntensity = _intensity;
 
-                half3 radiantIntensity = _pointLightColor * pointLightIntensity / (PIx4 * lightDistance * lightDistance);
+                float3 radiantIntensity = _pointLightColor * pointLightIntensity / (PIx4 * lightDistance * lightDistance);
 
-                half4 packedNormal = tex2D(_normalMap, i.uv);
+                float4 packedNormal = tex2D(_normalMap, i.uv);
                 float3 tangentNormal = UnpackNormal(packedNormal);
                 float3x3 tbn = float3x3(normalize(i.worldTangent), normalize(i.worldBitangent), normalize(i.worldNormal));
                 float3 worldNormal = normalize(mul(tangentNormal, tbn));
 
                 float NdotL01 = saturate(dot(worldNormal, lightVector));
                 float NdotV01 = saturate(dot(worldNormal, viewVector));
-                half3 surfaceIrradiance = radiantIntensity * NdotL01;
+                float3 surfaceIrradiance = radiantIntensity * NdotL01;
 
-                half3 albedo = tex2D(_albedo, i.uv).rgb;
-                half3 arm = tex2D(_arm, i.uv).rgb;
+                float3 albedo = tex2D(_albedo, i.uv).rgb;
+                float3 arm = tex2D(_arm, i.uv).rgb;
                 float ambientOcclusion = arm.r;
                 float roughness = arm.g;
                 float metallic = arm.b;
 
-                half3 BRDFLightFactor = microfacetBRDF(worldNormal, viewVector, lightVector, NdotV01, NdotL01, albedo.rgb, roughness, metallic);
-                half3 directLight = BRDFLightFactor * surfaceIrradiance;
+                float3 BRDFLightFactor = microfacetBRDF(worldNormal, viewVector, lightVector, NdotV01, NdotL01, albedo.rgb, roughness, metallic);
+                float3 directLight = BRDFLightFactor * surfaceIrradiance;
 
                 float3 indirectSpecularLight = SampleReflectionProbe(viewVector, worldNormal, roughness);
                 float3 indirectFresnel = fresnelReflectionWithSchlickApproximationAmbient(albedo, metallic, roughness, NdotV01);
                 float3 remainingAmbientDiffuseEnergy = 1.0 - indirectFresnel;
-                half3 diffuseAmbient = albedo * _ambientLightColor * remainingAmbientDiffuseEnergy * (1.0 - metallic);
-                half3 specularAmbient = indirectSpecularLight * indirectFresnel;
-                half3 ambientLight = (diffuseAmbient + specularAmbient) * ambientOcclusion;
+                float3 diffuseAmbient = albedo * _ambientLightColor * remainingAmbientDiffuseEnergy * (1.0 - metallic);
+                float3 specularAmbient = indirectSpecularLight * indirectFresnel;
+                float3 ambientLight = (diffuseAmbient + specularAmbient) * ambientOcclusion;
 
-                half3 surfaceLight = directLight + ambientLight;
+                float3 surfaceLight = directLight + ambientLight;
 
-                return half4 (surfaceLight, 1.0);
+                return float4 (surfaceLight, 1.0);
             }
             ENDCG
         }
