+++ Implementations/BuildingPerTheme/TexturesAndColor/PlanarProjectionOutsideSSAO.shader
@@ -163,7 +163,7 @@ Shader "ShaderCastle/Implementations/BuildingPerTheme/TexturesAndColor/PlanarPro
                 return saturate(finalVisibility); 
             }
 
-            half4 frag (v2f i) : SV_Target {
+            float4 frag (v2f i) : SV_Target {
                 // All vectors are normalized and point away from the surface
                 float3 viewVector = normalize(_WorldSpaceCameraPos - i.worldPos);
 
@@ -176,42 +176,42 @@ Shader "ShaderCastle/Implementations/BuildingPerTheme/TexturesAndColor/PlanarPro
                 }
                 
                 UNITY_LIGHT_ATTENUATION(attenuation, i, i.worldPos);
-                half3 radiantIntensity = _LightColor0.rgb * attenuation;
+                float3 radiantIntensity = _LightColor0.rgb * attenuation;
 
-                half4 packedNormal = tex2D(_normalMap, i.uv);
+                float4 packedNormal = tex2D(_normalMap, i.uv);
                 float3 tangentNormal = UnpackNormal(packedNormal);
                 float3x3 tbn = float3x3(normalize(i.worldTangent), normalize(i.worldBitangent), normalize(i.worldNormal));
                 float3 worldNormal = normalize(mul(tangentNormal, tbn));
 
                 float NdotL01 = saturate(dot(worldNormal, lightVector));
                 float NdotV01 = saturate(dot(worldNormal, viewVector));
-                half3 surfaceIrradiance = radiantIntensity * NdotL01;
+                float3 surfaceIrradiance = radiantIntensity * NdotL01;
 
-                half3 albedo = tex2D(_albedo, i.uv).rgb;
-                half3 arm = tex2D(_arm, i.uv).rgb;
+                float3 albedo = tex2D(_albedo, i.uv).rgb;
+                float3 arm = tex2D(_arm, i.uv).rgb;
                 float ambientOcclusion = arm.r;
                 float roughness = arm.g;
                 float metallic = arm.b;
 
-                half3 BRDFLightFactor = microfacetBRDF(worldNormal, viewVector, lightVector, NdotV01, NdotL01, albedo.rgb, roughness, metallic);
-                half3 directLight = BRDFLightFactor * surfaceIrradiance;
+                float3 BRDFLightFactor = microfacetBRDF(worldNormal, viewVector, lightVector, NdotV01, NdotL01, albedo.rgb, roughness, metallic);
+                float3 directLight = BRDFLightFactor * surfaceIrradiance;
 
-                half3 ambientLightColor = UNITY_LIGHTMODEL_AMBIENT.rgb;
+                float3 ambientLightColor = UNITY_LIGHTMODEL_AMBIENT.rgb;
                 float3 indirectSpecularLight = SampleReflectionProbe(viewVector, worldNormal, roughness);
                 float3 indirectFresnel = fresnelReflectionWithSchlickApproximationAmbient(albedo, metallic, roughness, NdotV01);
                 float3 remainingAmbientDiffuseEnergy = 1.0 - indirectFresnel;
-                half3 diffuseAmbient = albedo * ambientLightColor * remainingAmbientDiffuseEnergy * (1.0 - metallic);
-                half3 specularAmbient = indirectSpecularLight * indirectFresnel;
-                half3 ambientLight = (diffuseAmbient + specularAmbient) * ambientOcclusion;
+                float3 diffuseAmbient = albedo * ambientLightColor * remainingAmbientDiffuseEnergy * (1.0 - metallic);
+                float3 specularAmbient = indirectSpecularLight * indirectFresnel;
+                float3 ambientLight = (diffuseAmbient + specularAmbient) * ambientOcclusion;
 
                 UNITY_SETUP_STEREO_EYE_INDEX_POST_VERTEX(i);
                 float2 screenUV = i.screenPos.xy / i.screenPos.w;
                 float ambientOcclusionFactor = GetSSAO(screenUV, i.viewNormal); // 0 = fully occluded, 1 = unoccluded
                 ambientLight *= ambientOcclusionFactor;
 
-                half3 surfaceLight = directLight + ambientLight;
+                float3 surfaceLight = directLight + ambientLight;
 
-                return half4 (surfaceLight, 1.0);
+                return float4 (surfaceLight, 1.0);
             }
             ENDCG
         }
